+++ src/Exception/NotFoundException.php
@@ -5,8 +5,10 @@ namespace Http\Discovery\Exception;
 /**
  * Thrown when a discovery does not find any matches.
  *
+ * @final
+ *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class NotFoundException extends \RuntimeException
+/*final */class NotFoundException extends \RuntimeException
 {
 }